iOS: Move UTI display-state and intent types to their own file
