+++ iOS/DuckDuckGo/UnifiedToggleInput/UnifiedToggleInputCoordinator.swift
@@ -35,104 +35,6 @@ enum InputTextState {
     case userTyped
 }
 
-enum UnifiedToggleInputDisplayState: Equatable {
-    case hidden
-    case contextualChat
-    case aiTab(AITabState)
-    case omnibar(OmnibarState)
-
-    enum AITabState: Equatable {
-        case collapsed
-        case expanded
-    }
-
-    enum OmnibarState: Equatable {
-        case active
-        case inactive
-    }
-}
-
-enum UnifiedToggleInputIntent: Equatable {
-    case showCollapsed(from: UnifiedToggleInputDisplayState)
-    case showExpanded(from: UnifiedToggleInputDisplayState)
-    case showOmnibarEditing(expandedHeight: CGFloat, pendingExpandedHeight: CGFloat? = nil)
-    case showOmnibarInactive
-    case showOmnibarActive
-    case hideOmnibarEditing(animated: Bool)
-    case hide
-}
-
-/// First-class wrapper around "should this UI mutation animate, or snap?" Both branches
-/// take a closure of work to run — the wrapper handles the animation-context plumbing
-/// (`UIView.animate` vs `UIView.performWithoutAnimation`) so call sites stay flat.
-enum UTIAnimationStyle {
-    case snap
-    case animated(duration: TimeInterval, options: UIView.AnimationOptions, layoutTarget: UIView)
-
-    func perform(_ body: @escaping () -> Void, completion: ((Bool) -> Void)? = nil) {
-        switch self {
-        case .snap:
-            UIView.performWithoutAnimation(body)
-            completion?(true)
-        case let .animated(duration, options, layoutTarget):
-            UIView.animate(withDuration: duration, delay: 0, options: options, animations: {
-                body()
-                layoutTarget.layoutIfNeeded()
-            }, completion: completion)
-        }
-    }
-
-    /// Duration of this style — `0` for snap, the configured duration for animated. Useful
-    /// for matching parallel animations (e.g. `adjustUI(withKeyboardFrame:in:)`) to the same
-    /// timing.
-    var duration: TimeInterval {
-        switch self {
-        case .snap: return 0
-        case let .animated(duration, _, _): return duration
-        }
-    }
-}
-
-extension UnifiedToggleInputIntent {
-    enum AnimationConstants {
-        /// In-place pose morph on Duck.ai (collapsed ↔ expanded card).
-        static let aiTabPoseMorphDuration: TimeInterval = 0.35
-    }
-
-    /// Animation style derived from the intent's transition. The (from, to) pair is the only
-    /// signal needed — adding new transitions means adding new cases in the matrix below, not
-    /// scattering `if/else` across handlers.
-    func animationStyle(layoutTarget: UIView) -> UTIAnimationStyle {
-        switch self {
-        case let .showCollapsed(from):
-            return Self.transitionStyle(from: from, to: .aiTab(.collapsed), layoutTarget: layoutTarget)
-        case let .showExpanded(from):
-            return Self.transitionStyle(from: from, to: .aiTab(.expanded), layoutTarget: layoutTarget)
-        case .hide, .showOmnibarEditing, .showOmnibarInactive, .showOmnibarActive, .hideOmnibarEditing:
-            // `.hide` is always a snap (no animatable transition pair); the omnibar intents
-            // own bespoke animation flows and pick their own style.
-            return .snap
-        }
-    }
-
-    private static func transitionStyle(from: UnifiedToggleInputDisplayState,
-                                        to: UnifiedToggleInputDisplayState,
-                                        layoutTarget: UIView) -> UTIAnimationStyle {
-        switch (from, to) {
-        case (.aiTab(.expanded), .aiTab(.collapsed)),
-             (.aiTab(.collapsed), .aiTab(.expanded)):
-            return .animated(duration: AnimationConstants.aiTabPoseMorphDuration,
-                             options: .curveEaseInOut,
-                             layoutTarget: layoutTarget)
-        default:
-            // Fresh entry into a state from a different layout (tab swipe, app launch,
-            // omnibar dismiss, hide). The outer constraint changes shouldn't be animated;
-            // the destination should just BE there.
-            return .snap
-        }
-    }
-}
-
 enum ExternalSubmissionType {
     case query
     case prompt
